Clean up the translators a bit more of PhoSim and ImSim
This fixes some problems with only reading the primary header.
